♻️ [Refactor] 뷰컨트롤러에서 뷰 분리
